[test] Remove testStaticInit task, verify in main test task
The special task is there to test static initialization of a single
system property, which can be made more testable by extracting an
initializing method. This limits the risk that different tests that
reset the hooks would impact that staticInit test.

We fix that early on, but we already saw a manifestation of potential
issue on 3.3 (`master`):
 - JUnit 5 is used with both Jupiter engine and Vintage engine
 - a TestExecutionListener is added that resets all hooks after test
 methods but also classes and test engines
 - Gradle 6 seems to induce an empty execution of the Jupiter engine
 before the Vintage engine, which resets the hook (including the
 statically initialized GLOBAL_TRACE)