fix #2124 UnicastProcessor discards when onNext/cancel races
This is a follow-up to #2125, this time targeting the same class of
discarding race but on `UnicastProcessor`.

As there is a chance that the race references `actual` (to resolve the
discard hook) AFTER `cancel()`, we need to keep that field populated.
It was nulled out for the benefit of `downstreamCount()` (see #980) so
we replace that by a boolean, set to true when subscribed.

Reviewed-in: #2126