[Polish] Rework #2125 test cases into a single class
The abstract test class is replaced by a cross-cutting test class that
covers all relevant operators, which are provided as a collection of
DiscardLeakScenarios. This allows to also test multiple variants of an
operator without the need for one class per variant.

The tests have also been improved, with all hooks and schedulers setup
and teardown done in the test lifecycle methods.