fix #2143 Use ZoneId.of(String) to avoid blocking disk load
The former use of ZoneId.systemDefault() would trigger a file read from
disk (configuration read for the sytem's default timezone), despite the
timezone not being that important for the Scheduler. Using of(String)
factory method avoids that disk read. The ZoneId is frontloaded as a
constant anyway, since it is used by the SHUTDOWN constant as well.