Simplify Stacksafe logic and use a single static parallel scheduler...
Since operators are supposed to be non-blocking, switch to a parallel
scheduler that is shared among Stacksafe instances. The first time a
subscriber is actually wrapped, the scheduler is created.

This limits the number of potential threads created, and also there is
no more need for tracking number of protected subscribers since the
static scheduler shouldn't be shutdown on operators termination.

In this commit we also capture the maxStackDepth as a constructor
parameter, and we use 2 counters instead of a modulo to track the
need for wrapping. This allows to reset the counter when encountering a
publishOn or subscribeOn operator (which already resets the stack).