fix #2152 reportThrowInSubscribe triggers sneaky ClassCastException
This commit changes the Subscription set by reportThrowInSubscribe to
an instance compatible with QueueSubscription, preventing fused cases
where a ClassCastException would be thrown. The later would be
problematic with operators that make use of the Subscription when
receiving the onError signal, as said Subscription would be null at
that point due to the ClassCastException, sometimes causing hanging.

Reviewed-in: #2153