fix #2151 Allow publication of Gradle metadata, without optional deps
This commit:
 - removes propdeps plugin, known to now be problematic
 - uses a simplified local buildSrc plugin to provide the optional scope

Additionally it fixes building kotlin code by ensuring both reactor-core
and reactor-test declare an optional dependency to Kotlin stdlib.

The custom plugin should result in a Gradle module metadata file that
doesn't contain references to optional dependencies.

Last but not least, the JFrog Artifactory Gradle plugin has been bumped
to 4.15.2, as versions below 4.11 are known to prevent
distribution of the .module file to Artifactory.