Update Mono.java
Clarify Javadoc around `first()` and `or()` methods. IMHO the docs for the `first()` methods should be near enough identical as the only difference is the collection type (one being an array / varargs and the other being an iterator.) `or()` should explicitly refer to "signal" as `first()` does - referring to "result" instead implies it's not the first signal, but the first element emitted, which is confusing.