Introduce Sinks and deprecate concrete processors (#2188)
This commit introduces a large refactor of the Reactor processors.

Most of the time, such processors are not used as a Subscriber. Rather, they are
seen as a way to expose a Flux (or Mono) downstream while programmatically
feeding data to that downstream.

If done via direct calls of Subscriber methods, there is a need for external
synchronization (in order to comply with Reactive Streams specification) that is
often ignored by users.

The notion of `FluxProcessor#sink()` was originally added to help with that, but
getting the sink is still a second step, which can easily be forgotten.

So now we are deprecating the concrete implementations of FluxProcessor as well
as the `FluxProcessor#sink` methods. Instead, we introduce a top level utility
class to directly create standalone sinks, named `Sinks`.

As the concrete processors have been deprecated, the factory methods have been
mirrored in a `Processors` class. The only difference is that they appear to
return a generic `FluxProcessor` type (or rather, a new abstract subtype).

This will allow removal of the concrete implementations in 3.5.

This commit introduces a few additional interfaces around sinks:
 - SequenceSink and ScalarSink are a new base interface for sinks that have
 flux and mono semantics, respectively
 - StandaloneFluxSink and StandaloneMonoSink are the new interfaces exposed
 in `Sinks`
 - FluxSink and MonoSink are kept for operators purposes. Since each of such
 sinks naturally map to 1 Subscriber, they have additional methods to deal with
 request and cancellation that didn't make sense for standalone sinks.

It also introduces a new abstract class just between `FluxProcessor<A, B>`` and
concrete processors like `UnicastProcessor<T>`: `FluxIdentityProcessor<T>`.
This helps avoiding a transition from a ConcreteProcessor<T> to a
FluxProcessor<T,T>, which is more clunky in terms of generics.

The reference guide has been updated to reflect these changes around Processors
and put Sinks a bit more under the spotlight.

In order to cover some usage of specific getters in UnicastProcessor, it now
answers to `scan(ACTUAL)`.

Fixes #2179
Fixes #2203