Rework of M1 Sinks, further separate standalone Sinks API (#1037, #2218)
This commit is a rework of the `Sinks` approach introduced in 3.4.0-M1
(see #2179 and #2188).

Since the original goal is the ultimate hiding of most processors
implementations, the `Processors` factory has been removed. In a similar
effort NOT to give traction to processor use, `FluxIdentityProcessor`
has also been removed and replaced by the pre-existing equivalent of
`FluxProcessor<T, T>`.

The `StandaloneFluxSink` and `StandaloneMonoSink` have been reworked:
there is no `SequenceSink` and `ScalarSink` common superinterface, so
these new sinks don't share API with pre-existing `FluxSink` and
`MonoSink` anymore. The later are both reserved for in-operator sink
usage, with an API that reflects their association with a single
subscription.

Instead, this commit introduces flavors of sinks that are meant to be
used as publishers, regrouped under the `Sinks` class to clearly mark
they're 3 sides of a similar concept: `Sinks.Empty`, `Sinks.One` and
`Sinks.Many`. The intent is to always keep the `Sinks` prefix to clarify
the nature of the objects.

These sinks don't expose a fluent API, but rather expose `emitXxx`
methods that trigger relevant signals in their `Subscriber`s. They also
each expose a "view method", either `asMono()` or `asFlux`, which allows
downstream code to see the sink as a `Publisher` and subscribe to it.
These emit methods return a `Sinks.Emission` enum indicating success or
failure to emit, further distinguishing between cases for failure.
This later point fixes #1037.

This different API allows efficient instance creation since the
concrete implementations are processors which also directly implement
the sinks and their view.

Reviewed-in: #2218
Co-authored-by: Stephane Maldini <smaldini@netflix.com>
Co-authored-by: Simon Baslé <sbasle@vmware.com>