WIP #1526 Add doOnEach variant similar to doFirst&doOnCancel + Context
This commit adds doOnEachSubscriptionLifecycle, a dual to doOnEach
that allows listening to SUBSCRIBE, ON_SUBSCRIBE and CANCEL events in a
single lambda.

Additionally, contrary to doFirst/doOnCancel/doOnSubscribe it exposes
the Context.

Extending Signal with these new SignalType could have induced breaking
of existing code that doesn't strictly check the Signal's type when
processing it and using its getters. So this new operator doesn't expose
a Signal but works with a `BiConsumer<SignalType, Context>`. This is
assumed sufficient, as the "data" associated to such signals is of
limited use (basically Subscription for ON_SUBSCRIBE).