fix #2184 Add scope operator to limit range of Context manipulation
Provide an option (now Spec) per subscriber, don't expose ctx write
methods.

We expose one ScopeMutableSpec per Subscriber, which is passed to a
Consumer so that users can tune the scope options.

The definition of the scope is then done through a BiFunction that only
exposes the ContextView of both the original Context and the
scope-modified Context.

TODO: context loss detection for scope.