Decorate executors in start() rather than constructor (see #1778)
This commit delays the instantiation of backing ExecutorServices
in all vanilla Schedulers to the `start()` method, making it a
requirement to be called after construction.

This fixes a potentially problematic reference escape of `this`
from the constructors, since the executors are decorated using
a hook that takes the owning `Scheduler` as a second parameter.

Factory methods in `Schedulers` do automatically call `start()`
before returning a `Scheduler`.

As a side change, the ElasticScheduler no longer implicitly
start whenever a `Worker` is `pick()`ed, but has been made
compatible with restart (ie. start > dispose > start).
While disposed, tasks are rejected.