Rework specs to allow wrapping and put unsafe() at root level (#2418)
This commit reworks the Sinks specs so that the one() and empty() sinks
can have both a serialized and unsafe version.

Since  `unsafe()` lives at the ManySpec level, currently there's only
one set of `one()` and `empty()` factories.
So this commit introduces RootSpec and moves the `unsafe()` switch from
`ManySpec` to `Sinks` directly, retuning a `RootSpec` which allows
access to unsafe `one()` and `empty()`.

Internally, the implementation is also heavily simplified, collapsed
into only 2 internal classes for the specs.