Rework specs to allow wrapping and put unsafe() at root level (#2418)
This commit reworks the Sinks specs in preparation for two goals:
 - the one() and empty() should have a serialized version
 - further down the road we might need to introduce different wrapper(s)
 than the serializedManySink (eg. one that buffers in emit API to
 reorder work and perform work stealing on concurrent access)

The first goal implies that `Sinks.one()` is serialized and thus there
should be an unsafe version. But `unsafe()` lives at the ManySpec level.
So this commit introduces RootSpec and moves the `unsafe()` switch from
`ManySpec` to `Sinks` directly, retuning a `RootSpec` which allows
access to unsafe `one()` and `empty()`.

Internally, the implementation is turned into one that uses a common
`SinksWrapper` for all produced sinks, allowing to easily create custom
specs that will wrap the raw sinks with an arbitrary wrapper.

In order to avoid garbage, the vanilla wrappers are detected and static
instances of the various specs are returned instead of being created on
demand.