fix #2470 Make refguide BOM snippets attribute-based, bump via gradle
This commit adds a document attribute to the gettingStarted.adoc file
to store the BOM version displayed in Maven/Gradle snippets.

Additionally, the `bumpVersionsInReadme` Gradle task now copies that
file and perform variable substitution, replacing the previous version
with whichever BOM version is passed to the command line using
 `-PbomVersion=SOMETHING` (only affecting the attribute declaration).

If the attribute is not defined, no substitution should occur.