fix #2269 Add test util to capture logs in early-established loggers
This commit adds `LoggerUtils` to reactor-test. The main feature
is a method that installs a `Loggers` factory which wraps the current
factory and adds a level of indirection, to be used early in the
testbed lifecycle: `LoggerUtils.useCurrentLoggersWithCapture`.

Later on, in individual tests, one can enable log capture by passing
a `TestLogger` to `LoggerUtils.enableCaptureWith()` method.

All the logs will be copied to the `TestLogger` which allows
asserting not only loggers created since the beginning of the unit
test but also `Logger` instances that would have been instantiated
earlier (eg. in static block when classloading operators).

The only requirement is that these loggers have been created after
the call to `useCurrentLoggersWithCapture`.

Reviewed-in: #2492