[test] Have loggers reset to CapturingFactory after LoggersTest (#2512)
Depending on the execution order of the tests, the CapturingFactory set
up by the ReactorTestExecutionListener could be overwritten by
LoggersTest, resulting in failure of the 30 tests that rely on the log
capture mechanisms.

This commit ensures that the capturing factory is set again after these
tests, delegating to ReactorTestExecutionListener for the exact logger
setup.