Prevent nested install of CapturingFactory, polish LoggerUtils (#2512)
Now that logger factory is stored as a Function, the CapturingFactory
can reflexively detect another one is already in place (which was
trickier before, as the field would contain a method reference, not the
whole instance).

As a consequence, attempting to install a CapturingFactory over another
CapturingFactory will result in the old one being returned.

This commit also contains a few polish changes (fixing typos, etc...).