fix #2406 Amend SerializedSubscriberTest to accept rare double discard
As the fix potentially implies a non-trivial rewrite of FluxPublishOn,
we can improve things slightly by amending the test to detect double
discards and accept these.

That is only acceptable because the contract for discard hooks is that
they must be idempotent, even if no double discard would be preferable.