Extend RaceTestUtils to support N runnables (#2601)
Previously, RaceTestUtils.race() ran one of the runnables on the
provided scheduler and the other on the calling thread. Moreover,
to mimic support for N>2 runnables, users had to create a race()
hierarchy which defeated the purpose of synchronization in
practice.

After this change, some tests now fail consistently and 
exhibit race conditions with way less iterations. In particular,
a failure with singleOrEmpty has been found and will be addressed
via issue #2616.


Fixes #2593