Fix flatMap race favoring onComplete over onError (#2628)
This issue also affects the `merge` operator, since it relies on flatMap
internal classes. It would manifest in flatMap and merge seemingly
dropping errors and completing successfully.

The race is in the `innerError` method, which marks the inner as `done`
_before_ it updates the ERROR field with the inner error (or adds it to
the composite in case of delayErrors). This can cause a parrallel inner
onComplete signal to only observe the `inner.done` state without seeing
any `Throwable` in the ERROR field, which it interprets as a trigger
for downstream onComplete.

By ensuring the inner onError path only sets the `inner.done` to true
once the ERROR has been updated, we can avoid this race condition.

Fixes #2618.
Fixes #1941.