Prevent ordered merge hanging due to badly sized queue (#2631)
Prevent ordered merge hanging due to badly sized queue (#2631)ssssssssss
This commit ensures that the inner queues used by mergeOrdered operators
are congruent with the prefetch size. Since this is kind of an internal
requirement, the queue supplier isn't accepted through the constructors
anymore, but rather derived from the prefetch size in the inner
subscribers' initialization.

Previously, the supplier was actually ignored but the hardcoded supplier
wouldn't take the `prefetch` size into account, leading to data drops
and hanging when prefetch was larger than the queue size.

Fixes #1958.