Account for source disconnection in refCountGrace
This commit changes how FluxRefCountGrace evaluates if the connection
is terminated, by taking into account the case where the source
ConnectableFlux has itself disposed the connection (by checking if
the held `Disposable` that represents said connection is disposed).

This fixes a hanging situation when used in combination with publish.

Fixes #2649.