Avoid infinite discard loop on fused fromStream/fromIterable (#xxxx)
This commit reworks the `FluxIterableSubscription` and its conditional
counterpart to consider that the `QueueSubscription` `isEmpty()` in
the case where the subscription has been cancelled and the backing
iterator cannot safely be assumed to be finite.

This fixes an issue where fusing over these operators when wrapping an
infinite generator-style of `Iterable` or `Stream` would lead to an
infinit loop, because `poll()` would always produce a value and
`Operators.onDiscardQueueWithClear` performs its discard loop until
polling returns null.

Fixes #2654.