Fix both `Flux.single` variants when applied to a `Callable` (#xxxx)
This fixes two issues:
 - Flux.single() doesn't detect empty source if it is a Callable
 (excluding scalarCallables)
 - default value is currently not returned with the same source when
 using the single(T) variant

The reason is that in these cases Flux.single produces a `MonoCallable`
just like the Mono counterpart did before #2635 was fixed.

This commit ensures a `MonoSingleCallable` is returned, as introduced
in #2648.

In order to support `Flux.single(T)` (use case with a default value),
a relevant field is added to `MonoSingleCallable`.

Additionally, some tests have been moved to MonoSingleCallableTest and
the test coverage of MonoSingleCallable has been increased.
MonoSingleTest has a smoke test of the consistency of returned concrete
classes for all three types of sources, for all three variants of the
single operator/API (flux, flux with fallback, mono).