Polish MonoSingleCallableTest and MonoThenIgnoreTest
 - Removed FIXME in MonoSingleCallableTest

This will create conflict with master, since this FIXME was intended as
a reminder to add the scan(RUN_STYLE) test in master (which has now been
done).

 - Renamed MonoThenIgnoreTest to MonoIgnoreThenTest

This polishes the tests around Mono `then` family of operators: since
most tests pertain to the MonoIgnoreThen class, the test class is
renamed to reflect this.

Additionally, the few tests that don't actually rely on MonoIgnoreThen
but rather use alias `then()` are extracted in a Nested test set.