GitHub Actions Release Process:  Back to single publication (#2673)
This commit removes the extras publication, which caused issues
in the distribution of the pom.xml.

In order for the build file to be a bit more readable, the extra
artifact of docs.zip is added from `setup.gradle` rather than the
submodule's `reactor-core/build.gradle`. This implies checking
the project name and using `afterEvaluate` block, since setup is
applied to all submodules.

We also add a `-PforceSigning` option for debugging purposes.

Fixes #2672.