Polish corner cases in FluxLimitRequest
This commit ensures that several corner cases are now covered in the
FluxLimitRequest operator:

 - a negative cap is rejected at assembly time
 - a zero cap leads to immediate completion of the actual during
 subscription. the source isn't even subscribed to.
 - extraneous onNext/onError are detected and dropped

Note that for eg. `limitRequest(3)`, if the source emits exactly 3
elements then an `onError`, the `limitRequest` operator will still
consider itself `onComplete`d (and will now drop the error).

Similarly, `Flux.error(th).limitRequest(0)` will immediately _complete_
and the error won't be dropped in that case (since the error Flux won't
be subscribed to).

Fixes #2698.