[chore] Bump AssertJ to 3.19.0, polish tests and isolate tckTest (#2706)
AssertJ has introduced a lot of interesting new assertions, including
direct Stream assertions. On the other hand, bumping the version does
include a few breaking changes.

One of these introduces a conflict with the version of TestNG that we
use, because AssertJ now prioritize the TestNG class when attempting
to produce an _assumption_ exception, but the `SkipException` that
we have in scope at runtime is too old.

That drove the change of isolating TestNG dependency from the rest of
the tests, by separating the only part that uses it (the TCK tests)
in their own testSet.

Additional polishing of tests was made to suppress warnings and adapt
to changes in AssertJ 3.19.