Undeprecate Context#putAll(Context) and keep it for convenience
This commit reverts the deprecation of Context#putAll(Context) method,
which was slated for removal in 3.5.0 at the earliest.

It is instead kept as a convenience, eliminating the artificial warnings
caused by the deprecation in 3.4, especially in cases where one seeks
a short-hand way of adding several key-pairs to a `Context` (typically,
this is done via `context.putAll(Context.of(k1, v1, k2, v2, ...))`).

The added value is low, but the risk of keeping it is low too and the
compiler/JIT should be pretty efficient in simplifying that code.

Fixes #2718.