Add mergeComparing as fail-fast alternative to mergeOrdered (#2719)
This commit adds Flux.mergeComparing/mergeComparingWith as well as a
mergeComparingDelayError variant. The base operator fails fast if any
of its sources errors.

This supersedes mergeOrdered, which had the delayError behavior built in
(making it inconsistent with other merge operators). As a result, the
mergeOrdered operators are now deprecated.

Fixes #2640.

Co-authored-by: Simon Baslé <sbasle@vmware.com>