reworks FluxConcatArray to eliminate unexpected thread switch (#2742)
This commit replaces the WIP drain loop logic from the abstract
MultiSubscriptionSubscriber with a simpler guard based on ThreadLocal.
It works thanks to the serial nature of concatArray (onComplete triggers
the subscription to the next subscriber).

The old WIP guard and full drain loop weren't really necessary, other
than for detecting re-entrant `onComplete` triggered from the same
thread as the one that subscribed to the next publisher, and avoiding
stack overflows.

Thanks to ThreadLocals, we can detect such re-entrant case and let the
stack unwind, continuing subscription the next publisher inside the
original loop.

In the cases where `onComplete` is called from a different thread on the
other hand, that new thread must be favoured for starting the next
publisher. Once again, since we put the guard in a `ThreadLocal`, we
can ensure it is the case. The loop corresponding to the previous
publisher is "abandoned" and a new one is started for the new thread.

Fixes #2735.

Signed-off-by: Oleh Dokuka <odokuka@vmware.com>