Polish logging: test scenario name, verbosity of discard logs
This commit improves the verbosity of some logs in two categories.

1) Anonymous test scenarii dumping their class and line upon running:

This was confusing when looking at logs because it looked like part of
a stack trace. Now made clearer that this is a pseudo-name for a test
scenario that is missing a description.

The Scenario#description() method also makes it clear the stack element
returned is identifying a test scenario.

2) Operators now has a dedicated Logger for discard, and multidiscarding
uses the TRACE level:

As seen when running MonoCollectListTest, discarding a big collection
with a failing extractor or failing inner handler (the ones that are
invoked at least once per element in the collection) can lead to very
verbose logs.

All discarding logs are now depending on a dedicated Logger named
after the Operators.class + "-discard" suffix. Top level discarding
errors are logged at WARN level, but inner element discarding errors
are logged at TRACE level.

Additionally, two very similar messages in MonoCollectListTest were
amended to better reflect which test they originate from.