Polish logging: test scenario name, verbose MonoCollect logs (#2750)
This commit improves the verbosity of some logs in two categories.

1) Anonymous test scenarii dumping their class and line upon running:

This was confusing when looking at logs because it looked like part of
a stack trace. Now made clearer that this is a pseudo-name for a test
scenario that is missing a description.

The Scenario#description() method also makes it clear the stack element
returned is identifying a test scenario.

2) MonoCollect[List]Test have overly verbose logs

This is due to unnecessary throw in 4 tests, in the discard handlers.
Since these handlers are applied to each element in large collections,
one line of log per element is triggered.

The amended tests didn't seem to particularly rely on these throwing
handlers, so the exception throwing has been removed.