Prevent unnecessary duplicated lines in backtraces
This commit removes the `i` counter that was effectively guaranteeing
duplicated entries in the backtrace (eg. with a checkpoint) in the case
where the triggering exception was reused. The only differentiator was
then the indentation (driven by `i`).

This is especially visible when using a `checkpoint(String)` operator
along a `retry`. Provided the same exception instance is emitted on each
retry loop, `checkpoint` will add to the backtrace on each loop.

This commit removes the `i` from the tracking of existing backtraces and
thus ensures that for the same parent subscriber, prefix and description
there will be only one entry in the backtrace.

Fixes #2774.