Mono.create sink now discards if emitting after cancel (#2771)
Currently, contrary to flavors of sinks in `Flux.create`, when emitting
a value in the Mono.create sink there is no distinction between the
cancelled and terminated states. Any of these states leads to the value
being dropped (`Hooks.onNextDropped`).

This commit changes that and distinguish between terminated downstream
(`onNextDropped`) and cancelled (`Operators.onDiscard`).

Fixes #2769.