Use Named.of to describe parameterized tests input
This commit replaces various patterns introduced to pass arguments to
parameterized tests with a meaningful description of some sort (eg. a
`String type` extra parameter, or Object[][] arrangements).

The Arguments and Named classes can be leveraged to provide a meaningful
description to a parameter other than its toString, while using the
`{argumentsWithNames}` placeholder in the test description itself.

Fixes #2788.