Bump ByteBuddy to 1.11.16, use TestKit in agent integration test (#2776)
This commit bumps Byte-Buddy to 1.11.16, which makes the ad-hoc groovy
integration testing of the Byte Buddy gradle plugin feature non
functional due to breaking changes.

After much trial-and-error, the best way to fix that integration test
I found was to completely separate the test, as a inlined project in
resources. The entire project can then be copied in the `buildDir` and
executed using Gradle TestKit.

In order to get access to the snapshot of code under test, some token
replacement is used with the `files(@TOKEN@)` type of dependency.

This allowed to resurface the importance of using the `reactor-tools`
`-original.jar` version, as otherwise the ByteBuddy gradle plugin
will refuse to use the reactor Plugin (which implements the shaded
interface).

See https://github.com/raphw/byte-buddy/issues/833 for the original
issue that lead to the previous code.