Fix FluxSwitchMapNoPrefetch inner's scan(CANCELLED)
This commit fixes the scanUnsafe method of SwitchMapInner (the version
from FluxSwitchMapNoPrefetch). Currently, it returns an `int` for
an `Attr<Boolean>`.

Since that particular method was not tested, this commit duplicates the
tests found for `FluxSwitchMap` (and renames their test methods with the
`WithPrefetch` suffix).

Note that in a future version, `FluxSwitchMap` will be removed in favor
of `FluxSwitchMapNoPrefetch`.

Fixes #2807.